back/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose==3.3.0
bcrypt==4.1.2
email-validator==2.2.0
requests==2.32.3
python-multipart==0.0.9
pypdf==3.17.4